+++ packages/reef-knot/CHANGELOG.md
@@ -1,5 +1,48 @@
 # reef-knot
 
+## 7.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@7.0.1
+
+## 7.0.0
+
+### Major Changes
+
+- Metrics config refactored, replaced with certain actions callbacks
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@7.0.0
+  - @reef-knot/core-react@6.0.0
+  - @reef-knot/types@4.0.0
+  - @reef-knot/web3-react@6.0.0
+  - @reef-knot/wallets-list@4.0.0
+
+## 6.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@6.0.2
+
+## 6.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/connect-wallet-modal@6.0.1
+
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/ledger-connector@4.2.0
+
 ## 6.0.1
 
 ### Patch Changes