+++ packages/wallets/coin98/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @reef-knot/wallet-adapter-coin98
 
+## 4.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @reef-knot/types@4.0.0
+
 ## 3.0.0
 
 ### Patch Changes